Fix additional type mismatches reported by splint.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@3463 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/base64.h
@@ -12,7 +12,7 @@ NAME:
 
 #include "word.h"
 
-size_t base64_decode(word_t *word);
-bool   base64_validate(word_t *word);
+uint	base64_decode(word_t *word);
+bool	base64_validate(word_t *word);
 
 #endif	/* HAVE_BASE64_H */
